web_dsl/base/backend_base/requirements.txt
asyncio
commlib-py
pyyaml
websockets
fastapi
uvicorn
httpx
dotenv